WIP Commit around no-op constraints
This is an exploratory commit.

The basic idea investigated is to use a new type of deferred constraint which is
not enforced. The information is now storred in pg_index though and not in
pg_constraint which is not optimal. The rudimentary test case is the only one
implemented at the moment.