ALter Table Alter Column Type AOCO table optimization
Append-optimised column oriented (AOCO) tables are organised as one file per
column at storage layer.  This layout is amenable to optimisation in case of
ALTER TABLE ALTER COLUMN command where rewriting the entire table being altered
can be avoided.  The ADD COLUMN subcommand of ALTER TABLE is already optimised
like this for AOCO tables.

There is a catch. The file names on disk are semantically significant.

Bypass this by converting the ALTER COLUMN subcommand into DROP COLUMN and ADD
COLUMN subcommands during preparation step of ALTER TABLE.  The altered column
gets a new attribute number in catalog, new entry in pg_aoseg table and a new
file on disk.

This works great because the magic of MVCC can allow two versions of the
relation to exist as long as there are open transactions referring to them.

The current implementaiton is trying to limit the diversion with upstream and
some operation might have been fitted better in a different layer, though they
are grouped together to limit the foot print.

A bug was found that DROP column does not remove segment files from dropped
columns. This should be fixed in a different PR.

There are a couple TODOs:
	* deal with default values / constraints when coercing to the type.
According to the postgres documentation: `PostgreSQL will attempt to convert the
column's default value (if any) to the new type, as well as any constraints that
involve the column. But these conversions might fail, or might produce
surprising results. It's often best to drop any constraints on the column before
altering its type, and then add back suitably modified constraints afterwards.`
Maybe emiting a message that defaults will be silently dropped or aborting the
optimization is also fine...
 	* deal with existing indexes. Now the indexes are silently dropped. Same as
 above.