Alter Table Alter Column Set Data Type optimization
This is take two of the optimization for NOT re-writing the whole table when an
alter column set data type operation is performed on an AOCO table.

The first approach was splitting the command in two parts, DROP column and ADD
column. The major problem with this approach was the whole information about the
column has to be recreated, including defaults, constrains, comments etc. All
this logic would have to be replicated during the ADD column operation, which
also meant that the basic data structures involved would have to be enhanced in
order for the that information to be propagated forward.

In this commit, a rough sketch of another approach is coded. Add a
column_version information in the catalog for this column which will be
sequencially incremented on each Alter Column Set Data Type operation. The
VPINFO catalog entry is the ideal place for it. Of course a new XLOG record has
to be created for this operation which is missing in this commit.

Also on commit, the old file has to be cleared. On rollback, the new file has to
be cleared.

This commit is only a starting guide of the idea.

FormatAOSegmentFileName() is all lies. We can do without it honestly.